增加LoadGLTF示例工程
增加LoadGLTF示例工程，其中：
1. 调整了Pipeline.h头文件的依赖，减少了它对更多接口头文件的依赖情况
2. 增加了字符串路径替换的Utility方法
3. LoadGLTF示例工程尝试增加了以GLTF格式的模型资源加载，以及相关的Mesh Vertex Buffer的组织
4. LoadGLTF示例工程尝试向Shader增加更多的信息，以进一步增加c++对Shader的生成和编译功能的影响，更好地获取和调整Shader属性

+++ Core/D3D12/DX12Helpers.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "DX12Config.h"
 #include "../GraphicsInterface/Pipeline.h"
+#include "../GraphicsInterface/GPUResource.h"
 #include "../UVType.h"
 #include "../../Utility/InfoLog/InfoLog.h"
 #include <cassert>
@@ -439,10 +440,10 @@ inline D3D12_VIEWPORT ViewPortToDX12ViewPort(const ViewPort& input) {
 
 inline D3D12_RECT ScissorRectToDX12ScissorRect(const ScissorRect& input) {
 	D3D12_RECT output;
-	output.top = input.top;
-	output.left = input.left;
-	output.right = input.right;
-	output.bottom = input.bottom;
+	output.top = static_cast<LONG>(input.top);
+	output.left = static_cast<LONG>(input.left);
+	output.right = static_cast<LONG>(input.right);
+	output.bottom = static_cast<LONG>(input.bottom);
 	return output;
 }
